requirements.txt
accelerate==1.2.1
cycler==0.12.1
diffusers==0.31.0
einops==0.8.0
flash-attn==2.7.2.post1
fonttools==4.55.3
fsspec==2024.10.0
hjson==3.1.0
huggingface-hub==0.27.0
kiwisolver==1.4.7
matplotlib==3.10.0
numpy==2.0.1
peft==0.14.0
pyparsing==3.2.0
regex==2024.11.6
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
sympy==1.13.1
timm==1.0.12
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
scikit-image
torchmetrics
tensorboard
tensorboardX
pandas
deepspeed